Hibernation-info
Adds a specific safety-check on entering the Hibernation part of the
food meter and provided a quick hint as to the general idea.  Thanks for
the merge-cleanup and the guidance, Kevin.  Much appreciated.

Attempted to add memorial notes for engorgement, actually entering
hibernation, and successful exit; killed a successful hibernator and
just got the msglog and an empty killcount, though.  Not sure what I
missed.

Conflicts:
	src/player.cpp