Remove parameters from game::update_map.
The function basically initiates the map shifting and shifts all the monsters/npcs.
As it should always be called with (g->u.posx, g->u.posy) as parameters they can be implicitly assumed.
Also removed the special code for shifting while the player is a vehicle in map::shift.
Instead everything is done in game.cpp, but the call graph for moving while in a vehicle is this:
map::vehproc -> map::displace_vehicle -> game::update_map -> map::shift