Create new map caching infrastructure
Adds a per submap cache. That structure stores cached data for a single submap. It can be shifted with the submap and dose not need to be regenerated afterwards.

Moved the seen_cache into that structure.
Future plans for this:
Make transparency, outside and other cached things go there, too.
Add flags for what data is current and what needs to be regenerated.
After that one can add more things to the cache like move_cost.