Slightly enhanced what is drawn on the map
Only if the terrain is t_air, the terrain below will be shown.
This allows floor gratings to be shown, even when one can see through them.
Changed the second floor of the shelter (second floor) to have both new terrain types:
glass floor: see through, but can be walked and does not allow air through it.
floor grating: see through, allows air through, but can be walked on.