made zombies / monster attack across z-levels
Added a new flag MF_ATTACKS_UPWARDS for monster that can attack above their current z-level.
Monsters without that flag are restricted to attacks on current or directly below current z-level.
Monsters will not step into midair. That means the player is save if the monsters is on another z-level.