Move assignment of invlets to player class,
invlets are a a player thing, assignment must consider invlets of worn and wielded items anyway.

Separate it into two functions: one assigns a free invlet (or 0 if no invlet is available).
The other function forces an invlet (and takes one from the main inventory or the worn items if needed).
Both do nothing if called on a npc.

Remove the player parameter from inventory::restack. It is only used for invlet related code. This code uses the player (g->u) always.
The invlet related code is only executed when the inventory is the player's inventory, as npcs don't need invlets.

Remove some calls to inventory::assign_empty_invlet when not used with the parameter force set to true.
The inventory will assign an invlet automatically in inventory::add_item, so no need to do this before.
For displaying the invlet of the new item: player::i_add returns a reference to the added item, this reference has the correct invlet.