Implemented the item slot for container items:
This slot stores container content size (not item size), the watertight and the seals flag. Both are still present in item::flags, but not used from there.

Removed itype::is_container. Implemented inside of item class instead. Direct access to the item type is obsolete.

Implemented item::is_watertight to check for the presence of the two flags WATERTIGHT and SEALS. Used in player.cpp

Removed copying of the ammo type in vehicle.cpp - using a pointer to the ammo type is completely fine here.