Made tile cache work again. This time it's SDL2.
cata_tiles has now to local buffers: the previous buffer contains the last drawn frame.
When the view is scrolled (or the player simply moves a step) the content of that buffer (previous) is drawn (with an offset equal to the movement).
It is drawn to another buffer (current), than all the changed tiles are drawn to that buffer (including those that have not been visible before).
The current buffer is than drawn to the screen and the buffers are swapped.

tile_drawing_cache:
- added swap function
- made tile_type pointer and rotation value into a pair and use a single vector of those instead of two vectors
- removed operator= the compiler will add it
- made operator== simpler: the stdlib has an operator== for std::vector, use that one

cata_tiles:
- added comments to some functions and members
- removed get_tile_ratio(x|) and tile_ratio(x|y) as they are not used anymore
- rename screentile_(width|height) to screen_tile_count_(x|y) to make clear that it's the number of tiles, not of pixel.
- added a bunch of checks for the return values of SDL calls. Failures are logged using DebugLog.