Make game::look_around function with z-levels
One can stand on the roof of the evac shelter and look down to see the terrain below
Uses the up/down keys (2/8) a proper keybinding must be established!

Made the seen_cache function with z-levels - not nice at all, but it hides some things and shows others.
Made a primitive implementation of map::sees that respects z-levels - nothing nice or even good, but for now it functions as a placeholder for something better.
It uses the old version if source and target are on the same z-level.