Use oter_id::obj() instead of otermap
The result is the same, only now the look-up is done via oter_id::obj (which actually uses the integer id to look into the vector, which is faster than the string-based map look-up).

It also avoids the need to expose otermap, through it's still used.

Also note that it avoids repeated look-ups by storing a reference to the oter_t object and using that reference to get several properties of it.