Make generic_typed_reader work with types that don't have a comparator function
dding the dummy tempalte parameter makes those functions template functions. And those are only instanciated when they are actually used. If the FlagType does not have a `operator<`, the class `set<FlagType>` won't compile, similar if there is no `operator==`, `std::find` won't work.

If the functions *are* used, the compiler will complain, and the developer has to either add those operators as needed, or has to remove the call to the templated functions.