Remove inventory::add_stack and use push_back instead.
The content of the former function is moved into the later, all code that used to call the former now calls the later instead. This does not change the behavior.

Both functions were equivalent (as one did all the other anyway), and one of them is therefor redundant.
The function name `push_back` is used in the standard library and there is another function of that name already defined in the inventory class and it's used more often than `add_stack`.
This makes the name `push_back` better suited and it should therefor remain.