generate_bindings.lua: make overload resolution interface more flexible
This combines the last two parameters of the final function (which is called after the typed parameters have been collected from Lua) into one.
Instead of `cbc(a, b, c, d)` it is called as `cbc(a, b, { c, d })`.
In the future, the data object can be extended with even more data.