Update Lua bindings:
Remove overridden methods from Lua bindings. They are already exported as part of the parent class. A few functions were exported in the derived class, but they are actually defined in a base class and should be exported from there.

Fix return types of exported functions. This exports the correct return type instead of `nil` to Lua.

Remove useless exports to Lua:
- `get_var(, int)` does not exist, the wrapper would call the overload with a `double` as second parameter and that one is already exported.
- Calling `fill_with` requires a non-const string reference. The string   is actually modified in the function, but that is not returned to Lua. It is therefor quite useless.