gen: change retrieval of `this` object from Lua
Instead of having an explicit `instance` variable, which is loaded from the first object on the Lua stack, we treat it as a normal function parameter.
It is inserted into the overload resolution tree as first entry and is later loaded through `insert_overload_resolution` like any other parameter.
However, we  have to start the loading of the parameters from the first stack index instead of from the second (previously need to skip the already loaded `this` object). Therefor the change in the call to `insert_overload_resolution`.

We also need to cast the loaded parameter object as it is a proxy object, not a reference to the actually needed type. The proxy object provides a conversion operator that makes the cast work as expected.