Change the recipe_dictionary to contain values, not pointers
This gets rid of the need to delete the recipes manually. It's now properly done by the destructor of recipe_dictionary.

It also gets rid of the temporarily introduced references in the loops (over all recipes). The loop body can now use the loop variable directly, it's a reference now (and has the meaning of the removed variable).