Change vpart_info::sym to be a string.
This allows UTF8 encoded symbols to be used there.

The changed type is propagate through the various functions that used that value. They now use `std::string` instead of `char` or `long`.

A overload of `tileray::dir_symbol` has been added to allow it to be called with a string as parameter (and return type). The behavior is basically the same: it translates the first character of the string. Because the symbol was already defined as string in JSON (through the program would only load the first character of it), this does not change the outcome.

`special_symbol` has been changed to use `std::string`. The code for the vehicle part symbol is the only one that uses that function, so an additional overload for `long` is not needed.

Adds a bit of documentation to `special_symbol` and moves it from the source to the header file.
Link that function from documentation of `veh_type::symbol` to avoid repetition.

`special_symbol` returns Unicode box drawing symbols. Those are the same as referenced in the definition of the LINE_XXXX constants.