Change symbol of map_data_common_t to be a std::string
This allows UTF8 encoded symbols to be used there.

The changed type is propagate through some of the functions that used that value. They now use `std::string` instead of `char` or `long`.

Compatibility with old code that still requires / returns numbers is achieved by:
- using the first character of the UTF8 string,
- converting the character (given as number) to UTF8 via utf32_to_utf8.

Note that this is technically not correct for non-ASCII characters and will be fixed soon.