Call `craft_command::consume_components` unconditionally
even with "DEBUG_HS" - which is already handled by the function, it returns an empty list, there `used` will contain the same as if it had never been set.

Additionally: don't leave the function when the list is empty. The following code in the function can handle an empty list (having "DEBUG_HS" would result in one anyway).