Change game::clear_zombies to clear all creatures
Except the player character. It now removes npcs and monsters alike.

It is only called once in the game itself (and there it is followed by clearing the `game::active_npc` vector), so changing that place to call `clear_creatures` does not change the behaviour.

The other places that call it are in the test code. Those may as well clear the npcs too.