Make CBM items normal tools.
Instead of having a specific case in player::use for CBM item, it will handle them as tools, call their iuse function, which is now a simply iuse_actor that calls the bionic installation function.
The installation difficulty can be specified in the iuse actor.
The actor can install any bionic from any item, the items id is not used as bionic id anymore.

This erase the islot_bionic type.