Allow resizing the terminal and re-initialize the ui when that happens.
Currently only working from the top level of the main menu screen and from the main game loop. And only in ncurses-builds.

Moves some code in `main_menu::opening_screen` into a lambda, so it can be called again later.
The `WINDOW_PTR` need to be on the stack of `opening_screen`, not on the stack of the lambda.