Move saving the generated submaps into map::loadn
`map::generate` is called here and in "editmap.cpp".
In "editmap.cpp" it's called on a temporary map, which deletes the contained submap pointers on its own. They must therefor *not* be stored.

For `map::loadn` nothing has effectively changed. The same code is run, only in the caller now.