Move content of game::shared_from into Creature_tracker function.
The `game` function now merely forwards to the `Creature_tracker` function.
Therefor also moves the documentation of that function over there.

Changes `Creature_tracker::player_character` to be a `shared_ptr` as that allows it to be returned directly.
(It's much better than the previous hack of creating a `shared_ptr` with an empty deleter function.)