Combine game::load_npcs and game::unload_npcs and game::reload_npcs
Only need one function to rule them all.

The function skips already loaded npcs, unloads npcs that are no longer in the reality bubble and makes npcs active that are now in the reality bubble.

It also fixes the TODO: `on_unload` is only invoked (via `Creature_tracker::despawn`) for npcs that leave the reality bubble. `on_load` is only invoked for npcs that became active.