Change rename_file to use new path class
And rename it to just `rename` as that's what it's called in the standard C++ library version.

Also move it into the `cata` namespace for consistency with the other functions.

Also changed the implementation by moving the common parts of the function (the header and the closing `}`) out of the conditional compilation. They are (and in fact *must*) be the same for every build type.