Lua: ensure wrapper files are not re-generated on every run
Lua has an effectively random ordering of contents in a table, so iterating over a table will yield a different sequence on every run of the script.
This means the output file contain the definitions of the various things in random order.
And the file is always different to the previously generated file.

This triggers recompilation far to often with no actual need.