Change Lua wrapper to work via a library
Changes the Makefile to link all Lua related code into a single library file and later link that into the final executable.

For non-Lua builds, a separate file implements all wrapper functions (that are used by the main game code) with dummy behaviour. This file is made into the library and linked later.