gen: change how the class name is forwarded in overload resolution
Instead of forwarding it as function parameter, it is now part of the overload resolution data.

This enables correct casting of the `this` object:
```C++
class Parent {
	void func();
};
class Child : public Parent {
	void func(int);
};
```

The generated wrapper must cast the `Child` object to `Parent` when calling `func` without any parameter.
Calling `child_instance.func()` won't work as `Child::func` requires a parameter. Casting it to `Parent` first avoids this: `static_cast<Parent&>(child_instance).func()` works fine.

Note that this does not interfere with calling virtual functions: calling a virtual function always invokes the proper dispatch as generated by the C++ compiler. It does not matter whether the functions is called on a `Child` or on a `Parent` object, as long as both are virtual.