Lua/lex: write generated wrapper code into several files.
This just makes compilation simpler and faster. Instead of compiling one big cpp file, we can not compile several smaller files.
It also means when an exported class has changed, don't need to re-compile code that has no reference to the changed class, as it will be in other cpp files.