Move vehicle::remove_item(index) into item_stack
So it can be used by other code as well (e.g. via `map_stack`).
Rename it to just `remove` as it's quite obvious that it removes an item.
Changed the parameter to be a `size_t` as it's an index.
Removed the return value and changed the behaviour to *require* a valid index.