Change map::board_vehicle to check things on its own
If there is no part to board, nothing is done.

Take a player as reference (it's mandatory).

Don't change the position of the player (that must be done by the caller) and we don't need to call `game::update_map`.
Note that *all* existing calls used the position of the creature as argument anyway, so this "feature" is not used anyway.