Change vehicle::install_part to not check whether it can be installed
Instead it installs it always. The caller is responsible for checking before this.

Note: calls with a `vehicle_part` would always succeed (e.g. in "iuse.cpp") so no change there.
Other code ignored the return value anyway or just assumed it to be valid (e.g. "veh_utils.cpp").

Code in "savegame_json.cpp" requires a check now as it relied on `install_part` to dismiss installing an already existing part.

Code in "veh_type.cpp" will now display the reason why the part can not be mounted.

The function now also returns the installed part as reference.